renamed delivery to delivery-service in model and js files
