+++ app/models/point.js
@@ -5,7 +5,7 @@ const Point = new Schema({
 	latitude: { type: String, required: true },
 	longitude: { type: String, required: true },
 	reference: { type: String },
-	delivery: { type: Schema.Types.ObjectId, ref: 'delivery' }
+	deliveryService: { type: Schema.Types.ObjectId, ref: 'delivery_services' }
 }, { collection: 'points', timestamps: true });
 
 module.exports = mongoose.model('points', Point);
